+++ nimbus-jose-jwt_aws-kms-extension/build.gradle
@@ -8,7 +8,7 @@ plugins {
 }
 
 group = 'software.amazon.lynx'
-version = '1.0.1'
+version = '1.1.0'
 
 java {
     toolchain {
