+++ nimbus-jose-jwt_aws-kms-extension/src/main/java/com/nimbusds/jose/aws/kms/crypto/KmsAsymmetricRSASSAVerifier.java
@@ -18,106 +18,32 @@ package com.nimbusds.jose.aws.kms.crypto;
 
 
 import com.amazonaws.services.kms.AWSKMS;
-import com.amazonaws.services.kms.model.DependencyTimeoutException;
-import com.amazonaws.services.kms.model.DisabledException;
-import com.amazonaws.services.kms.model.InvalidGrantTokenException;
-import com.amazonaws.services.kms.model.InvalidKeyUsageException;
-import com.amazonaws.services.kms.model.KMSInternalException;
-import com.amazonaws.services.kms.model.KMSInvalidSignatureException;
-import com.amazonaws.services.kms.model.KMSInvalidStateException;
-import com.amazonaws.services.kms.model.KeyUnavailableException;
 import com.amazonaws.services.kms.model.MessageType;
-import com.amazonaws.services.kms.model.NotFoundException;
-import com.amazonaws.services.kms.model.VerifyRequest;
-import com.amazonaws.services.kms.model.VerifyResult;
-import com.nimbusds.jose.CriticalHeaderParamsAware;
-import com.nimbusds.jose.JOSEException;
-import com.nimbusds.jose.JWSHeader;
-import com.nimbusds.jose.JWSVerifier;
-import com.nimbusds.jose.RemoteKeySourceException;
 import com.nimbusds.jose.aws.kms.crypto.impl.KmsAsymmetricRSASSAProvider;
-import com.nimbusds.jose.aws.kms.exceptions.TemporaryJOSEException;
-import com.nimbusds.jose.crypto.impl.CriticalHeaderParamsDeferral;
-import com.nimbusds.jose.util.Base64URL;
-import java.nio.ByteBuffer;
 import java.util.Set;
 import javax.annotation.concurrent.ThreadSafe;
 import lombok.NonNull;
-import lombok.var;
 
 /**
  * Sign verifier implementation for RSA-SSA signing with public/private key stored in AWS KMS.
  * <p>
  * See {@link KmsAsymmetricRSASSAProvider} for supported algorithms, and for details of various
  * constructor parameters.
+ *
+ * @deprecated use {@link KmsAsymmetricVerifier} instead.
  */
 @ThreadSafe
-public class KmsAsymmetricRSASSAVerifier
-        extends KmsAsymmetricRSASSAProvider
-        implements JWSVerifier, CriticalHeaderParamsAware {
-
-    /**
-     * The critical header policy.
-     */
-    private final CriticalHeaderParamsDeferral critPolicy = new CriticalHeaderParamsDeferral();
-
+@Deprecated
+public class KmsAsymmetricRSASSAVerifier extends KmsAsymmetricVerifier {
 
     public KmsAsymmetricRSASSAVerifier(
             @NonNull final AWSKMS kms, @NonNull final String privateKeyId, @NonNull final MessageType messageType) {
         super(kms, privateKeyId, messageType);
     }
 
-
     public KmsAsymmetricRSASSAVerifier(
             @NonNull final AWSKMS kms, @NonNull String privateKeyId, @NonNull final MessageType messageType,
             @NonNull final Set<String> defCritHeaders) {
-        super(kms, privateKeyId, messageType);
-        critPolicy.setDeferredCriticalHeaderParams(defCritHeaders);
-    }
-
-
-    @Override
-    public Set<String> getProcessedCriticalHeaderParams() {
-
-        return critPolicy.getProcessedCriticalHeaderParams();
-    }
-
-
-    @Override
-    public Set<String> getDeferredCriticalHeaderParams() {
-
-        return critPolicy.getDeferredCriticalHeaderParams();
-    }
-
-
-    @Override
-    public boolean verify(
-            @NonNull final JWSHeader header, @NonNull final byte[] signedContent, @NonNull final Base64URL signature)
-            throws JOSEException {
-
-        if (!critPolicy.headerPasses(header)) {
-            return false;
-        }
-
-        var message = getMessage(header, signedContent);
-
-        VerifyResult verifyResult;
-        try {
-            verifyResult = getKms().verify(new VerifyRequest()
-                    .withKeyId(getPrivateKeyId())
-                    .withSigningAlgorithm(header.getAlgorithm().toString())
-                    .withMessageType(getMessageType())
-                    .withMessage(message)
-                    .withSignature(ByteBuffer.wrap(signature.decode())));
-        } catch (KMSInvalidSignatureException e) {
-            return false;
-        } catch (NotFoundException | DisabledException | KeyUnavailableException | InvalidKeyUsageException
-                | KMSInvalidStateException e) {
-            throw new RemoteKeySourceException("An exception was thrown from KMS due to invalid key.", e);
-        } catch (DependencyTimeoutException | InvalidGrantTokenException | KMSInternalException e) {
-            throw new TemporaryJOSEException("A temporary exception was thrown from KMS.", e);
-        }
-
-        return verifyResult.isSignatureValid();
+        super(kms, privateKeyId, messageType, defCritHeaders);
     }
 }
