+++ nimbus-jose-jwt_aws-kms-extension/src/main/java/com/nimbusds/jose/aws/kms/crypto/impl/KmsAsymmetricRSASSAProvider.java
@@ -41,82 +41,13 @@ import org.apache.commons.codec.digest.MessageDigestAlgorithms;
 
 /**
  * This class provides cryptography support for RSA-SSA based signing/verification with keys stored in AWS KMS.
+ *
+ * @deprecated use {@link KmsAsymmetricSigningCryptoProvider} instead.
  */
-public abstract class KmsAsymmetricRSASSAProvider extends BaseJWSProvider {
-
-    /**
-     * AWS-KMS client.
-     */
-    @NonNull
-    @Getter(AccessLevel.PROTECTED)
-    private final AWSKMS kms;
-
-    /**
-     * KMS private-key (CMK) ID (it can be a key ID, key ARN, key alias or key alias ARN)
-     */
-    @NonNull
-    @Getter(AccessLevel.PROTECTED)
-    private final String privateKeyId;
-
-    /**
-     * KMS Message Type. Refer KMS's sign and verify APIs for details.
-     * Ref: https://docs.aws.amazon.com/kms/latest/APIReference/API_Sign.html#KMS-Sign-request-MessageType
-     */
-    @NonNull
-    @Getter(AccessLevel.PROTECTED)
-    private final MessageType messageType;
-
-    public static final Map<JWSAlgorithm, String> JWS_ALGORITHM_TO_MESSAGE_DIGEST_ALGORITHM =
-            ImmutableMap.<JWSAlgorithm, String>builder()
-                    .put(JWSAlgorithm.parse(SigningAlgorithmSpec.RSASSA_PKCS1_V1_5_SHA_256.toString()),
-                            MessageDigestAlgorithms.SHA_256)
-                    .put(JWSAlgorithm.parse(SigningAlgorithmSpec.RSASSA_PKCS1_V1_5_SHA_384.toString()),
-                            MessageDigestAlgorithms.SHA_384)
-                    .put(JWSAlgorithm.parse(SigningAlgorithmSpec.RSASSA_PKCS1_V1_5_SHA_512.toString()),
-                            MessageDigestAlgorithms.SHA_512)
-                    .put(JWSAlgorithm.parse(SigningAlgorithmSpec.RSASSA_PSS_SHA_256.toString()),
-                            MessageDigestAlgorithms.SHA_256)
-                    .put(JWSAlgorithm.parse(SigningAlgorithmSpec.RSASSA_PSS_SHA_384.toString()),
-                            MessageDigestAlgorithms.SHA_384)
-                    .put(JWSAlgorithm.parse(SigningAlgorithmSpec.RSASSA_PSS_SHA_512.toString()),
-                            MessageDigestAlgorithms.SHA_512)
-                    .build();
-
-    /**
-     * The supported JWS algorithms (alg) by the RSA-SSA provider class.
-     *
-     * Note: We are using KMS prescribed algorithm names here.
-     * Ref: https://docs.aws.amazon.com/kms/latest/developerguide/symm-asymm-choose.html#key-spec-rsa-sign
-     */
-    public static final Set<JWSAlgorithm> SUPPORTED_ALGORITHMS = JWS_ALGORITHM_TO_MESSAGE_DIGEST_ALGORITHM.keySet();
-
+@Deprecated
+public abstract class KmsAsymmetricRSASSAProvider extends KmsAsymmetricSigningCryptoProvider {
     protected KmsAsymmetricRSASSAProvider(
             @NonNull final AWSKMS kms, @NonNull final String privateKeyId, @NonNull final MessageType messageType) {
-        super(SUPPORTED_ALGORITHMS);
-        this.kms = kms;
-        this.privateKeyId = privateKeyId;
-        this.messageType = messageType;
-    }
-
-    protected ByteBuffer getMessage(final JWSHeader header, final byte[] signingInput) throws JOSEException {
-        final var alg = header.getAlgorithm();
-        var message = signingInput;
-
-        String messageDigestAlgorithm = Optional.ofNullable(JWS_ALGORITHM_TO_MESSAGE_DIGEST_ALGORITHM.get(alg))
-                .orElseThrow(() -> new JOSEException(
-                        String.format("No digest algorithm exist for JWE algorithm %s in map: %s",
-                                alg, JWS_ALGORITHM_TO_MESSAGE_DIGEST_ALGORITHM)));
-
-        if (messageType == MessageType.DIGEST) {
-            MessageDigest messageDigestProvider;
-            try {
-                messageDigestProvider = MessageDigest.getInstance(messageDigestAlgorithm);
-            } catch (NoSuchAlgorithmException e) {
-                throw new JOSEException("Invalid message digest algorithm.", e);
-            }
-            message = messageDigestProvider.digest(message);
-        }
-
-        return ByteBuffer.wrap(message);
+        super(kms, privateKeyId, messageType);
     }
 }
